refactor(og): extract asset ownership policy
